fix: change to write client for mutation
