+++ src/dataService/index.ts
@@ -1,2 +1,3 @@
 export * from './savedItemsService';
 export * from './tagDataService';
+export * from './usersMetaService';
