--- src/dataService/mutationServices/index.ts
@@ -1,3 +1,0 @@
-export * from './tagsMutationService';
-export * from './savedItemsMutationService';
-export * from './usersMetaService';
