+++ src/dataService/savedItemsService.ts
@@ -1,7 +1,7 @@
 import { Knex } from 'knex';
 import { knexPaginator as paginate } from '@pocket-tools/apollo-cursor-pagination';
-import { IContext } from '../../server/context';
-import { cleanAndValidateTag, mysqlTimeString } from '../utils';
+import { IContext } from '../server/context';
+import { cleanAndValidateTag, mysqlTimeString } from './utils';
 import {
   Pagination,
   SavedItem,
@@ -11,8 +11,10 @@ import {
   SavedItemsFilter,
   SavedItemsSort,
   SavedItemStatus,
-} from '../../types';
-import config from '../../config';
+  SavedItemUpsertInput,
+} from '../types';
+import config from '../config';
+import { ItemResponse } from '../externalCaller/parserCaller';
 
 type DbResult = {
   user_id?: number;
@@ -37,8 +39,11 @@ export class SavedItemDataService {
     [SavedItemStatus.DELETED]: 'DELETED',
     [SavedItemStatus.HIDDEN]: 'HIDDEN',
   };
-  private readDb: Knex;
+  private db: Knex;
+  private writeDb: Knex;
   private readonly userId: string;
+  private readonly apiId: string;
+
   private sortMap = {
     CREATED_AT: '_createdAt',
     UPDATED_AT: '_updatedAt',
@@ -47,8 +52,10 @@ export class SavedItemDataService {
   };
 
   constructor(context: IContext) {
-    this.readDb = context.db.readClient;
+    this.db = context.db.readClient;
+    this.writeDb = context.db.writeClient;
     this.userId = context.userId;
+    this.apiId = context.apiId;
   }
 
   public static convertDbResultStatus(dbResult: DbResult): DbResult;
@@ -82,28 +89,28 @@ export class SavedItemDataService {
    * For now just to reuse the same query and reduce testing burden :)
    */
   public buildQuery(): any {
-    return this.readDb('list').select(
+    return this.db('list').select(
       'given_url AS url',
       'item_id AS id',
       'resolved_id AS resolvedId', // for determining if an item is pending
       'favorite as isFavorite',
-      this.readDb.raw(
+      this.db.raw(
         'CASE WHEN favorite = 1 THEN UNIX_TIMESTAMP(time_favorited) ELSE null END as favoritedAt '
       ),
       'time_favorited', // for pagination sort
       'status',
-      this.readDb.raw(
+      this.db.raw(
         `CASE WHEN status = ${SavedItemStatus.ARCHIVED} THEN true ELSE false END as isArchived`
       ),
-      this.readDb.raw('UNIX_TIMESTAMP(time_added) as _createdAt'),
+      this.db.raw('UNIX_TIMESTAMP(time_added) as _createdAt'),
       'time_added', // for pagination sort
       'item_id',
-      this.readDb.raw('UNIX_TIMESTAMP(time_updated) as _updatedAt'),
+      this.db.raw('UNIX_TIMESTAMP(time_updated) as _updatedAt'),
       'time_updated', // for pagination sort
-      this.readDb.raw(
+      this.db.raw(
         `CASE WHEN status = ${SavedItemStatus.DELETED} THEN UNIX_TIMESTAMP(time_updated) ELSE null END as _deletedAt`
       ),
-      this.readDb.raw(
+      this.db.raw(
         `CASE WHEN status = ${SavedItemStatus.ARCHIVED} THEN UNIX_TIMESTAMP(time_updated) ELSE null END as archivedAt`
       )
     );
@@ -173,7 +180,7 @@ export class SavedItemDataService {
     }
     return await paginate(
       // Need to use a subquery in order to order by derived fields ('archivedAt')
-      this.readDb.select('*').from(query.as('page_query')),
+      this.db.select('*').from(query.as('page_query')),
       {
         first: pagination?.first,
         last: pagination?.last,
@@ -252,8 +259,8 @@ export class SavedItemDataService {
    * @param itemId
    */
   public async getSavedItemTimeRead(itemId: string): Promise<any> {
-    return this.readDb('list')
-      .select(this.readDb.raw('SQL_NO_CACHE time_read'))
+    return this.db('list')
+      .select(this.db.raw('SQL_NO_CACHE time_read'))
       .where({ item_id: itemId, user_id: this.userId })
       .first();
   }
@@ -323,7 +330,7 @@ export class SavedItemDataService {
     baseQuery: Knex,
     isHighlighted: boolean
   ): Promise<any> {
-    const highlightSubquery = this.readDb('user_annotations')
+    const highlightSubquery = this.db('user_annotations')
       .select('user_id as hl_user_id', 'item_id as hl_item_id')
       .where('user_id', this.userId)
       .andWhere('status', 1)
@@ -347,7 +354,7 @@ export class SavedItemDataService {
             'list.item_id'
           );
         })
-        .andWhere(this.readDb.raw('highlights.hl_item_id is null'));
+        .andWhere(this.db.raw('highlights.hl_item_id is null'));
     }
   }
 
@@ -390,7 +397,7 @@ export class SavedItemDataService {
       return baseQuery;
     }
 
-    const tagsSubQuery = this.readDb('item_tags')
+    const tagsSubQuery = this.db('item_tags')
       .select(
         'tag as tag_tag',
         'user_id as tag_user_id',
@@ -422,6 +429,214 @@ export class SavedItemDataService {
     }
     // Tags are a many-to-one relationship with item, so need
     // to take distinct results after performing this join
-    return this.readDb.select('*').from(baseQuery.as('base')).distinct();
+    return this.db.select('*').from(baseQuery.as('base')).distinct();
+  }
+
+  /**
+   * Format a date to the configured database timezone
+   * @param date
+   * @private
+   */
+  private static formatDate(date: Date): string {
+    return mysqlTimeString(date, config.database.tz);
+  }
+
+  /**
+   * Update the 'favorite' attribute of an item, and the auditing fields
+   * in the table ('time_updated', etc.)
+   * @param itemId the item ID to update
+   * @param favorite whether the item is a favorite or not
+   */
+  public async updateSavedItemFavoriteProperty(
+    itemId: string,
+    favorite: boolean
+  ): Promise<void> {
+    this.db = this.writeDb;
+
+    const timestamp = SavedItemDataService.formatDate(new Date());
+    const timeFavorited = favorite ? timestamp : '0000-00-00 00:00:00';
+    await this.writeDb('list')
+      .update({
+        favorite: +favorite,
+        time_favorited: timeFavorited,
+        time_updated: timestamp,
+        api_id_updated: this.apiId,
+      })
+      .where({ item_id: itemId, user_id: this.userId });
+  }
+
+  /**
+   * Update the 'status' attribute of an item, for archive/unarchive,
+   * and the auditing fields in the table ('time_updated', etc.)
+   * @param itemId the item ID to update
+   * @param archived whether the item is a favorite or not
+   */
+  public async updateSavedItemArchiveProperty(
+    itemId: string,
+    archived: boolean
+  ): Promise<void> {
+    this.db = this.writeDb;
+
+    const timestamp = SavedItemDataService.formatDate(new Date());
+    const timeArchived = archived ? timestamp : '0000-00-00 00:00:00';
+    const status = archived ? 1 : 0;
+    // TODO: Do we care if this makes an update that doesn't change the status?
+    // e.g. archiving an already archived item will update
+    //    time_read, time_upated, api_id_updated; but not status
+    await this.writeDb('list')
+      .update({
+        status: status,
+        time_read: timeArchived,
+        time_updated: timestamp,
+        api_id_updated: this.apiId,
+      })
+      .where({ item_id: itemId, user_id: this.userId });
+  }
+
+  /**
+   * Soft delete a saved item. Since we delete from multiple tables,
+   * we perform the entire operation as a single transaction
+   * to allow us to fully rollback should any on of the
+   * database statements fail.
+   * @param itemId
+   */
+  public async deleteSavedItem(itemId) {
+    this.db = this.writeDb;
+
+    const transaction = await this.writeDb.transaction();
+    try {
+      // remove tags for saved item
+      await transaction('item_tags').delete().where({
+        user_id: this.userId,
+        item_id: itemId,
+      });
+
+      // remove attribution for saved item
+      await transaction('item_attribution').delete().where({
+        user_id: this.userId,
+        item_id: itemId,
+      });
+
+      // remove scroll position for saved item
+      await transaction('items_scroll').delete().where({
+        user_id: this.userId,
+        item_id: itemId,
+      });
+
+      // update status for saved item to soft delete
+      await transaction('list')
+        .update({
+          status: SavedItemStatus.DELETED,
+          time_updated: SavedItemDataService.formatDate(new Date()),
+          api_id_updated: this.apiId,
+        })
+        .where({ item_id: itemId, user_id: this.userId });
+
+      await transaction.commit();
+    } catch (err) {
+      await transaction.rollback();
+      throw err;
+    }
+  }
+
+  /**
+   * Undelete a saved item. Check the time_read for the saved item to determine
+   * which status the item is assigned when moved from the deleted state.
+   * @param itemId
+   */
+  public async updateSavedItemUnDelete(itemId) {
+    this.db = this.writeDb;
+
+    const query: any = this.getSavedItemTimeRead(itemId);
+
+    // This is a check to determine if the saved item was previously archived. Fun, right?
+    const status =
+      query.time_read === '0000-00-00 00:00:00'
+        ? SavedItemStatus.UNREAD
+        : SavedItemStatus.ARCHIVED;
+
+    await this.writeDb('list')
+      .update({
+        status,
+        time_updated: SavedItemDataService.formatDate(new Date()),
+        api_id_updated: this.apiId,
+      })
+      .where({ item_id: itemId, user_id: this.userId });
+  }
+
+  /**
+   * @param item
+   * @param savedItemUpsertInput
+   */
+  public async upsertSavedItem(
+    item: ItemResponse,
+    savedItemUpsertInput: SavedItemUpsertInput
+  ): Promise<any> {
+    this.db = this.writeDb;
+
+    const currentDate = SavedItemDataService.formatDate(new Date());
+    const givenTimestamp = new Date(savedItemUpsertInput.timestamp * 1000);
+    const givenDate = savedItemUpsertInput.timestamp
+      ? SavedItemDataService.formatDate(givenTimestamp)
+      : currentDate;
+    //`returning` is not supported for mysql in knex
+    await this.writeDb('list')
+      .insert({
+        user_id: parseInt(this.userId),
+        item_id: item.itemId,
+        given_url: savedItemUpsertInput.url,
+        status: 0,
+        resolved_id: item.resolvedId,
+        title: item.title,
+        time_added: givenDate,
+        time_updated: currentDate,
+        time_read: '0000-00-00 00:00:00',
+        time_favorited: savedItemUpsertInput.isFavorite
+          ? givenDate
+          : '0000-00-00 00:00:00',
+        favorite: savedItemUpsertInput.isFavorite ? 1 : 0,
+        api_id: parseInt(this.apiId),
+        api_id_updated: parseInt(this.apiId),
+      })
+      .onConflict()
+      .merge();
+  }
+
+  /**
+   * Build a query to update the `time_updated` field of many items
+   * the list table, by item id.
+   * @param itemIds The item IDS to update the `time_Updated` to now.
+   */
+  public updateListItemMany(itemIds: string[]): Knex.QueryBuilder {
+    return this.listItemUpdateBuilder().whereIn('item_id', itemIds);
+  }
+
+  /**
+   * Build a query to update the `time_updated` field of one item in
+   * the list table, by item id.
+   * @param itemId
+   */
+  public updateListItemOne(itemId: string): Knex.QueryBuilder {
+    return this.listItemUpdateBuilder().where('item_id', itemId);
+  }
+
+  /**
+   * Helper function to build updates to a user's list.
+   * Used to mark updates that affect the list item (e.g. a new tag
+   * association) but are not direct updates to the list table.
+   * Does not include the necessary `join` or `where` statement
+   * to properly execute this query.
+   * Do not run this query as-is. Should only be used to compose other
+   * queries. That's why it's private :)
+   */
+  private listItemUpdateBuilder(): Knex.QueryBuilder {
+    this.db = this.writeDb;
+
+    return this.writeDb('list')
+      .update({
+        time_updated: SavedItemDataService.formatDate(new Date()),
+        api_id_updated: this.apiId,
+      })
+      .andWhere('user_id', this.userId);
   }
 }
