+++ src/dataService/tagDataService.spec.ts
@@ -1,7 +1,7 @@
 import chai, { expect } from 'chai';
-import { TagMutationService } from '../mutationServices';
-import { TagCreateInput } from '../../types';
+import { TagCreateInput } from '../types';
 import deepEqualInAnyOrder from 'deep-equal-in-any-order';
+import { TagDataService } from './index';
 chai.use(deepEqualInAnyOrder);
 
 describe('deduplicateInput', () => {
@@ -11,7 +11,7 @@ describe('deduplicateInput', () => {
       { name: 'foam', savedItemId: '1' },
       { name: 'roller', savedItemId: '2' },
     ];
-    const deduplicated = TagMutationService.deduplicateTagInput(inputData);
+    const deduplicated = TagDataService.deduplicateTagInput(inputData);
     expect(deduplicated.length).to.equal(2);
     expect(deduplicated).to.deep.equalInAnyOrder(inputData.slice(1));
   });
@@ -20,7 +20,7 @@ describe('deduplicateInput', () => {
       { name: 'foam', savedItemId: '1' },
       { name: 'foamy', savedItemId: '1' },
     ];
-    const deduplicated = TagMutationService.deduplicateTagInput(inputData);
+    const deduplicated = TagDataService.deduplicateTagInput(inputData);
     expect(deduplicated.length).to.equal(2);
     expect(deduplicated).to.deep.equalInAnyOrder(inputData);
   });
