+++ src/test/functional/mutationServices.test/tagsMutationService-create.integration.ts
@@ -16,7 +16,8 @@ import { getUnixTimestamp } from '../../../utils';
 chai.use(deepEqualInAnyOrder);
 
 describe('Mutation for Tag: ', () => {
-  const db = readClient();
+  //using write client as mutation will use write client to read as well.
+  const db = writeClient();
   const eventEmitter: ItemsEventEmitter = new ItemsEventEmitter();
   const server = new ApolloServer({
     schema: buildFederatedSchema({ typeDefs, resolvers }),
@@ -59,7 +60,7 @@ describe('Mutation for Tag: ', () => {
   let clock;
 
   afterAll(async () => {
-    await db.destroy();
+    await readClient().destroy();
     await writeClient().destroy();
     clock.restore();
   });
